requirements.txt
Django==4.0.5
mysqlclient==2.0.3
djangorestframework==3.12.2
django-debug-toolbar==3.4.0
django-filter==22.1